Commit to Test Quiz 6

+++ main.h
@@ -10,7 +10,7 @@ int _putchar(char c);
 int _printf(const char *format, ...);
 char *_itoa(int n);
 char *_uitoa(unsigned int n);
-char *_uitoa_base(unsigned int n, unsigned int base);
+char *_uitoa_base(unsigned long n, unsigned int base);
 int print_int_helper(int n);
 int print_int(va_list args);
 char *conv_to_binary(unsigned int num);
